# cbt-software/ARCHITECTURE/openapi-runtime.yaml
openapi: 3.0.0
paths:
...
  /reports/overall-performance:
    get:
      operationId: ReportsController_overallPerformance
      parameters:
        - name: schoolId
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: &ref_15
        - Reports
  /reports/question-difficulty:
    get:
      operationId: ReportsController_questionDifficulty
      parameters:
        - name: schoolId
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_15
  /reports/student-results/{classId}:
    get:
      operationId: ReportsController_getStudentResults
      parameters:
        - name: classId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_15
  /reports/generate-student-result/{studentId}/{classId}:
    post:
      operationId: ReportsController_generateStudentResult
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            type: string
        - name: classId
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      tags: *ref_15
  /reports/student-result/{id}:
    get:
      operationId: ReportsController_getStudentResult
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_15
    put:
      operationId: ReportsController_updateStudentResult
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_15
    delete:
      operationId: ReportsController_deleteStudentResult
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_15
  /reports/generate-report-card/{studentId}/{classId}:
    post:
      operationId: ReportsController_generateReportCard
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            type: string
        - name: classId
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      tags: *ref_15
  /reports/report-card/{id}:
    get:
      operationId: ReportsController_getReportCard
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_15
    put:
      operationId: ReportsController_updateReportCard
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_15
  /reports/report-cards/{classId}:
    get:
      operationId: ReportsController_getReportCards
      parameters:
        - name: classId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_15
  /reports/report-card/{id}/publish:
    put:
      operationId: ReportsController_publishReportCard
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_15
...
info:
  title: CBT Platform API
  description: Runtime generated OpenAPI document
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...